requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
GDAL==3.5.3
Markdown==3.4.1
postgis==1.0.4
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
